+++ portal/www/portal/slice-add-resources-jacks.js
@@ -32,6 +32,12 @@ function fileupload_onchange()
     // change RSpec dropdown menu back to 'Choose RSpec'
     clear_other_inputs('#file_select');
 
+    // Enable/disable the auxiliary 'select' button depending on the rspec file
+    // console.log("URF = " + user_rspec_file);
+    if(user_rspec_file)
+	$('#reapply_file_select').prop('disabled', false);
+    else
+	$('#reapply_file_select').prop('disabled', true);
 
 }
 
@@ -254,6 +260,9 @@ function rspec_onchange()
     // If RSPEC is unchosen, disable reserve resources
     if (rspec_opt == "") {
 	disable_reserve_resources();
+	$('#reapply_rspec_select').prop("disabled", true);
+    } else {
+	$('#reapply_rspec_select').prop("disabled", false);
     }
 }
 
@@ -339,6 +348,9 @@ function jacks_fetch_topology_callback(rspecs) {
   } else if (jacksEditorApp.submittingRspec) {
       jacksEditorApp.submittingRspec = false;
       validateSubmit();
+  } else if (jacksEditorApp.appendingRspec) {
+      jacksEditorApp.appendingRspec = false;
+      jacksEditorApp.appendToTopology(rspec);
   } else if (jacksEditorApp.passingContextToURL != null) {
       var editor_url = jacksEditorApp.passingContextToURL;
       jacksEditorApp.passingContextToURL = null;
@@ -347,8 +359,8 @@ function jacks_fetch_topology_callback(rspecs) {
       jacksEditorApp.invoking_global_node = false;
       do_global_node_addition_internal();
   } else if (jacksEditorApp.invoking_auto_ip) {
-      jacksEditorApp.invoking_auto_ip = false;
       do_auto_ip_assignment_internal();
+      jacksEditorApp.invoking_auto_ip = false;
   } else if (jacksEditorApp.invoking_selection_duplicate_nolinks) {
       jacksEditorApp.invoking_selection_duplicate_nolinks = false;
       do_selection_duplicate_internal(false);