+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 # [Release 3.11](https://github.com/GENI-NSF/geni-portal/milestones/3.11)
 
+* Remove Flack
+  ([#1676](https://github.com/GENI-NSF/geni-portal/issues/1676))
 
 # [Release 3.10](https://github.com/GENI-NSF/geni-portal/milestones/3.10)